refactor(mcp): split report runtime helpers
