+++ crates/codelens-mcp/src/tools/session/metrics_config.rs
@@ -1,9 +1,10 @@
+use crate::AppState;
 use crate::protocol::BackendKind;
 use crate::session_metrics_payload::build_session_metrics_payload;
 use crate::tool_defs::{
-    default_budget_for_preset, default_budget_for_profile, ToolPreset, ToolProfile, ToolSurface,
+    ToolPreset, ToolProfile, ToolSurface, default_budget_for_preset, default_budget_for_profile,
 };
-use crate::tools::{success_meta, AppState, ToolResult};
+use crate::tool_runtime::{ToolResult, success_meta};
 use serde_json::json;
 
 pub fn get_watch_status(state: &AppState, _arguments: &serde_json::Value) -> ToolResult {
